# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido criado com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar um novo pedido",
        "tags": [
          "orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de pedidos retornada com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos os pedidos",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/start-tracking": {
      "post": {
        "operationId": "OrdersController_startTracking",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rastreamento iniciado com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Iniciar o rastreamento de uma ordem",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/end-tracking/{id}": {
      "post": {
        "operationId": "OrdersController_endTracking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do rastreamento",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rastreamento finalizado com sucesso."
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderTracking"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Finalizar o rastreamento de uma ordem",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/report/{id}": {
      "get": {
        "operationId": "OrdersController_getOrderReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da ordem",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Relatório retornado com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter relatório de uma ordem",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/etapas": {
      "post": {
        "operationId": "OrdersController_createEtapa",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da ordem",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "example": "Corte"
                  },
                  "funcionarioCode": {
                    "type": "string",
                    "example": "FUNC001"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Etapa criada com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar uma nova etapa para uma ordem",
        "tags": [
          "orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_listEtapasByOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da ordem",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Etapas retornadas com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar etapas de uma ordem",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/etapas/{id}/iniciar": {
      "post": {
        "operationId": "OrdersController_startEtapa",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da etapa",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Etapa iniciada com sucesso."
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Etapa"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Iniciar uma etapa",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/etapas/{id}/finalizar": {
      "post": {
        "operationId": "OrdersController_endEtapa",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da etapa",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Etapa finalizada com sucesso."
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Etapa"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Finalizar uma etapa",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/atualizar-status": {
      "post": {
        "operationId": "OrdersController_atualizarStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID da ordem",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "em_andamento"
                  },
                  "motivoId": {
                    "type": "number",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status atualizado com sucesso."
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar o status de uma ordem",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do pedido retornados com sucesso."
          },
          "404": {
            "description": "Pedido não encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhes de um pedido por ID",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/motivos-interrupcao": {
      "get": {
        "operationId": "OrdersController_listMotivosInterrupcao",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de motivos de interrupção retornada com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos os motivos de interrupção",
        "tags": [
          "orders"
        ]
      },
      "post": {
        "operationId": "OrdersController_createMotivoInterrupcao",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMotivoInterrupcaoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Motivo de interrupção criado com sucesso."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Adicionar um novo motivo de interrupção",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/historico-producao": {
      "post": {
        "operationId": "OrdersController_createHistoricoProducao",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHistoricoProducaoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registro de histórico de produção criado com sucesso."
          },
          "404": {
            "description": "Pedido, funcionário ou motivo de interrupção não encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar um novo registro no histórico de produção",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/historico-producao/{id}": {
      "put": {
        "operationId": "OrdersController_updateHistoricoProducao",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do registro de histórico de produção",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHistoricoProducaoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registro de histórico de produção atualizado com sucesso."
          },
          "404": {
            "description": "Registro de histórico de produção, pedido, funcionário ou motivo de interrupção não encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar um registro no histórico de produção",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/historico-producao": {
      "get": {
        "operationId": "OrdersController_listHistoricoProducao",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico de produção retornado com sucesso."
          },
          "404": {
            "description": "Pedido não encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar histórico de produção de um pedido",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}/rastreamentos": {
      "get": {
        "operationId": "OrdersController_listRastreamentosByOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rastreamentos retornados com sucesso."
          },
          "404": {
            "description": "Ordem não encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos os rastreamentos de uma ordem",
        "tags": [
          "orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS API",
    "description": "API documentation for NestJS with Swagger, Orval, and MikroORM",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "productCode": {
            "type": "string",
            "description": "Código do produto",
            "example": "PROD001"
          },
          "employeeCode": {
            "type": "string",
            "description": "Código do funcionário",
            "example": "FUNC001"
          },
          "maquinaCodigo": {
            "type": "string",
            "description": "Código da máquina designada (opcional)",
            "example": "MAQ-001"
          },
          "lotQuantity": {
            "type": "number",
            "description": "Quantidade do lote",
            "example": 100
          },
          "finalDestination": {
            "type": "string",
            "description": "Destino final do lote",
            "example": "Setor de Montagem"
          }
        },
        "required": [
          "productCode",
          "employeeCode",
          "lotQuantity",
          "finalDestination"
        ]
      },
      "TrackOrderDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "number",
            "description": "ID da ordem de serviço",
            "example": 1
          },
          "employeeCode": {
            "type": "string",
            "description": "Código do funcionário",
            "example": "FUNC001"
          },
          "lostQuantity": {
            "type": "number",
            "description": "Quantidade de peças perdidas",
            "example": 5
          },
          "processedQuantity": {
            "type": "number",
            "description": "Quantidade de peças processadas",
            "example": 100
          }
        },
        "required": [
          "orderId",
          "employeeCode"
        ]
      },
...
      "Funcionario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "code": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "cargo": {
            "type": "string"
          },
          "salario": {
            "type": "number"
          },
          "password": {
            "type": "string"
          },
          "auth": {
            "$ref": "#/components/schemas/Auth"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "default": "2025-05-20T22:05:55.117Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "default": "2025-05-20T22:05:55.117Z"
          }
        },
        "required": [
          "id",
          "code",
          "nome",
          "cargo",
          "salario",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "OrderTracking": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID único do rastreamento",
            "example": 1
          },
          "order": {
            "description": "Pedido associado",
            "allOf": [
              {
                "$ref": "#/components/schemas/Order"
              }
            ]
          },
          "funcionarios": {
            "description": "Funcionário responsável",
            "allOf": [
              {
                "$ref": "#/components/schemas/Funcionario"
              }
            ]
          },
          "startTime": {
            "format": "date-time",
            "type": "string",
            "description": "Hora de início do rastreamento",
            "example": "2023-10-01T12:00:00Z"
          },
          "endTime": {
            "format": "date-time",
            "type": "string",
            "description": "Hora de término do rastreamento",
            "example": "2023-10-01T14:00:00Z"
          },
          "lostQuantity": {
            "type": "number",
            "description": "Quantidade de peças perdidas",
            "example": 5
          },
          "processedQuantity": {
            "type": "number",
            "description": "Quantidade de peças processadas",
            "example": 100
          }
        },
        "required": [
          "id",
          "order",
          "funcionarios",
          "startTime"
        ]
      },
      "Etapa": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID único da etapa",
            "example": 1
          },
          "nome": {
            "type": "string",
            "description": "Nome da etapa",
            "example": "Corte"
          },
          "order": {
            "description": "Pedido associado",
            "allOf": [
              {
                "$ref": "#/components/schemas/Order"
              }
            ]
          },
          "funcionario": {
            "description": "Funcionário responsável",
            "allOf": [
              {
                "$ref": "#/components/schemas/Funcionario"
              }
            ]
          },
          "inicio": {
            "format": "date-time",
            "type": "string",
            "description": "Horário de início da etapa",
            "example": "2023-10-01T12:00:00Z"
          },
          "fim": {
            "format": "date-time",
            "type": "string",
            "description": "Horário de término da etapa",
            "example": "2023-10-01T14:00:00Z"
          }
        },
        "required": [
          "id",
          "nome",
          "order",
          "funcionario"
        ]
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID único do pedido",
            "example": 1
          },
          "orderNumber": {
            "type": "string",
            "description": "Número da Ordem de Pedido (OP)",
            "example": "OP-12345"
          },
          "name": {
            "type": "string",
            "description": "Nome do pedido",
            "example": "Pedido de cortes especiais"
          },
          "product": {
            "description": "Produto associado",
            "allOf": [
              {
                "$ref": "#/components/schemas/Product"
              }
            ]
          },
          "funcionarioResposavel": {
            "description": "Funcionário responsável",
            "allOf": [
              {
                "$ref": "#/components/schemas/Funcionario"
              }
            ]
          },
          "maquina": {
            "description": "Máquina designada para o serviço",
            "allOf": [
              {
                "$ref": "#/components/schemas/Maquina"
              }
            ]
          },
          "lotQuantity": {
            "type": "number",
            "description": "Quantidade de peças no lote",
            "example": 100
          },
          "finalDestination": {
            "type": "string",
            "description": "Destino final do lote",
            "example": "Setor de Montagem"
          },
          "trackings": {
            "default": [],
            "description": "Rastreamentos do pedido",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderTracking"
            }
          },
          "status": {
            "type": "string",
            "default": "aberto",
            "description": "Status do pedido",
            "example": "aberto"
          },
          "etapas": {
            "default": [],
            "description": "Etapas do pedido",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Etapa"
            }
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "default": "2025-05-20T22:05:55.107Z",
            "description": "Data de criação do pedido",
            "example": "2023-10-01T12:00:00Z"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "default": "2025-05-20T22:05:55.107Z",
            "description": "Data de atualização do pedido",
            "example": "2023-10-01T14:00:00Z"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "name",
          "product",
          "funcionarioResposavel",
          "lotQuantity",
          "finalDestination",
          "trackings",
          "status",
          "etapas",
          "created_at",
          "updated_at"
        ]
      },
      "CreateMotivoInterrupcaoDto": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "string",
            "description": "Código do motivo de interrupção",
            "example": "M001"
          },
          "descricao": {
            "type": "string",
            "description": "Descrição do motivo de interrupção",
            "example": "Falta de material"
          }
        },
        "required": [
          "codigo",
          "descricao"
        ]
      },
      "CreateHistoricoProducaoDto": {
        "type": "object",
        "properties": {
          "pedidoId": {
            "type": "number",
            "description": "ID do pedido associado",
            "example": 1
          },
          "funcionarioId": {
            "type": "number",
            "description": "ID do funcionário responsável",
            "example": 1
          },
          "acao": {
            "type": "string",
            "description": "Ação realizada",
            "example": "início"
          },
          "detalhes": {
            "type": "string",
            "description": "Detalhes adicionais sobre a ação",
            "example": "Início da etapa de corte"
          },
          "motivoInterrupcaoId": {
            "type": "number",
            "description": "ID do motivo de interrupção",
            "example": 1
          },
          "dataHora": {
            "format": "date-time",
            "type": "string",
            "description": "Data e hora em que a ação foi registrada",
            "example": "2023-10-01T12:00:00Z"
          }
        },
        "required": [
          "pedidoId",
          "funcionarioId",
          "acao"
        ]
      },
...
    }
  }
}